otu_abundance() function
This is what I used in my M.S. to created relative abundance graphs instead of the plot_bar() function included in the phyloseq package. It has an easy way to decide which taxa of a given rank to include in the graph; all remaining taxa will be grouped together as "Other" so you don't have to end up with 70 different phyla in your graph.